hrtimers: Split out hrtimer_types.h
mainline inclusion
from mainline-v6.8-rc1
category: performace

We need to reduce the scope of what's included in sched.h: task_struct
includes a hrtimer, so split out the core types into their own header.

Cc: Thomas Gleixner <tglx@linutronix.de>
Conflicts:
	include/linux/hrtimer.h
(cherry picked from commit 50d91c76582513852e38eb80491f54d44cfb51fc)
